Removed a hack that worked around an incorrect return type on gadget.DigitalProxy.getData().
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@14301 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ vrj.net/examples/CSharp/SimpleApp/SimpleApp.cs
@@ -74,12 +74,12 @@ public class SimpleApp : vrj.GlApp
    {
       Console.WriteLine("SimpleApp.preFrame()");
 
-      if ( mButton0.getProxy().getData() != (int) gadget.Digital.State.OFF )
+      if ( mButton0.getProxy().getData() != gadget.Digital.State.OFF )
       {
          Console.WriteLine("Button 0 pressed");
       }
 
-      if ( mButton1.getProxy().getData() != (int) gadget.Digital.State.OFF )
+      if ( mButton1.getProxy().getData() != gadget.Digital.State.OFF )
       {
          Console.WriteLine("Button 1 pressed");
       }
